Restructuring of api calls
