+++ src/pages/index.vue
@@ -3,13 +3,13 @@ import Button from 'primevue/button';
 import InputText from 'primevue/inputtext';
 import Dropdown from 'primevue/dropdown';
 
+import SearchBox from '@/components/search/SearchBox.vue';
+
 
 import { ref } from 'vue'
 
 import 'primeicons/primeicons.css'
-import axios from 'axios';
 
-const searchParam = ref('');
 const selectedNum = ref<string | number>();
 const pageSizes = [
     25,
@@ -18,31 +18,14 @@ const pageSizes = [
     'all'
 ]
 
-const searchByName = async () => {
-    try {
-        const { data } = await axios.get(`https://pokeapi.co/api/v2/pokemon/${searchParam.value}`);
-        if (data.name) {
-            alert('success!')
-        }
-        console.log(data);
-    } catch (error) {
-        alert('No pokemon found with that name! Try again')
-        throw error
-    }
 
-}
 
 
 </script>
 
 <template>
     <div class="card-container main-box">
-        <div class="search-form p-card">
-            <label for="username">Pokemon search</label>
-            <InputText @keyup.enter="searchByName" v-model="searchParam" id="username" aria-describedby="username-help" />
-            <small id="username-help">Enter the complete name of a pokemon please</small>
-            <Button @click="searchByName">Search</Button>
-        </div>
+      <search-box></search-box>
         <div class="dropdown p-card card">
             <!-- <h3>Pokemon display</h3> -->
             <Dropdown v-model="selectedNum" :options="pageSizes" placeholder="How many pokemon per page?" />
